requirements.txt
weaviate-client
openai
tqdm
python-dotenv
